refactor: 메소드 중복 제거,helper method public으로 변경
