+++ src/test/java/subway/line/LineAcceptanceTestHelper.java
@@ -6,7 +6,6 @@ import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.HashMap;
-import java.util.Map;
 import org.springframework.http.MediaType;
 
 public class LineAcceptanceTestHelper {
@@ -30,7 +29,7 @@ public class LineAcceptanceTestHelper {
         return params;
     }
 
-    static HashMap<String, String> 노선_파라미터_생성2() {
+    public static HashMap<String, String> 노선_파라미터_생성2() {
         HashMap<String, String> params = new HashMap<>();
         params.put("name", "2호선");
         params.put("color", "bg-green-600");
@@ -40,23 +39,7 @@ public class LineAcceptanceTestHelper {
         return params;
     }
 
-    public static ExtractableResponse<Response> 지하철역_조회_요청() {
-        return RestAssured.given().log().all()
-                          .when().get("/stations")
-                          .then().log().all()
-                          .extract();
-    }
-
-    public static ExtractableResponse<Response> 지하철역_생성_요청(Map<String, String> params) {
-        return RestAssured.given().log().all()
-                          .body(params)
-                          .contentType(MediaType.APPLICATION_JSON_VALUE)
-                          .when().post("/stations")
-                          .then().log().all()
-                          .extract();
-    }
-
-    static ExtractableResponse<Response> 노선목록_조회_요청() {
+    public static ExtractableResponse<Response> 노선목록_조회_요청() {
         return RestAssured.given().log().all()
                           .contentType(MediaType.APPLICATION_JSON_VALUE)
                           .when().get("/lines")
@@ -64,7 +47,7 @@ public class LineAcceptanceTestHelper {
                           .extract();
     }
 
-    static ExtractableResponse<Response> 노선_단건조회_요청(ExtractableResponse<Response> response) {
+    public static ExtractableResponse<Response> 노선_단건조회_요청(ExtractableResponse<Response> response) {
         return RestAssured.given().log().all()
                           .contentType(MediaType.APPLICATION_JSON_VALUE)
                           .when().get("/lines/" + getIdFromResponse(response))
@@ -72,7 +55,7 @@ public class LineAcceptanceTestHelper {
                           .extract();
     }
 
-    static ExtractableResponse<Response> 노선_수정_요청(HashMap<String, String> updateParam,
+    public static ExtractableResponse<Response> 노선_수정_요청(HashMap<String, String> updateParam,
         Long id) {
         return RestAssured.given().log().all()
                           .body(updateParam)
@@ -82,14 +65,14 @@ public class LineAcceptanceTestHelper {
                           .extract();
     }
 
-    static HashMap<String, String> 노선수정_파라미터_생성() {
+    public static HashMap<String, String> 노선수정_파라미터_생성() {
         HashMap<String, String> updateParam = new HashMap<>();
         updateParam.put("name", "다른분당선");
         updateParam.put("color", "bg-red-600");
         return updateParam;
     }
 
-    static ExtractableResponse<Response> 노선_삭제_요청(Long id) {
+    public static ExtractableResponse<Response> 노선_삭제_요청(Long id) {
         return RestAssured.given().log().all()
                           .contentType(MediaType.APPLICATION_JSON_VALUE)
                           .when().delete("/lines/" + id)
