+++ src/test/java/subway/station/StationAcceptanceTestHelper.java
@@ -9,13 +9,13 @@ import org.springframework.http.MediaType;
 
 public class StationAcceptanceTestHelper {
 
-    static Map<String, String> 지하철_파라미터_생성(String name) {
+    public static Map<String, String> 지하철_파라미터_생성(String name) {
         Map<String, String> params = new HashMap<>();
         params.put("name", name);
         return params;
     }
 
-    static ExtractableResponse<Response> 지하철역_조회_요청() {
+    public static ExtractableResponse<Response> 지하철역_조회_요청() {
         return RestAssured.given().log().all()
                           .contentType(MediaType.APPLICATION_JSON_VALUE)
                           .when().get("/stations")
@@ -23,10 +23,19 @@ public class StationAcceptanceTestHelper {
                           .extract();
     }
 
-    static ExtractableResponse<Response> 지하철역_제거_요청(Long id) {
+    public static ExtractableResponse<Response> 지하철역_제거_요청(Long id) {
         return RestAssured.given().log().all()
                           .contentType(MediaType.APPLICATION_JSON_VALUE)
                           .when().delete("/stations/" + id)
+                          .then().log().all()
+                          .extract();
+    }
+
+    public static ExtractableResponse<Response> 지하철역_생성_요청(Map<String, String> params) {
+        return RestAssured.given().log().all()
+                          .body(params)
+                          .contentType(MediaType.APPLICATION_JSON_VALUE)
+                          .when().post("/stations")
                           .then().log().all()
                           .extract();
     }
